requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
decorator==4.4.2
idna==3.10
imageio==2.31.2
imageio-ffmpeg==0.5.1
moviepy==1.0.3
numpy>=1.21
Pillow>=10.0.0,<11.0.0
proglog==0.1.10
requests==2.31.0
tqdm==4.66.6
urllib3==2.0.7
